remove dependancy on ledger for common types
