+++ applications/minotari_ledger_wallet/comms/Cargo.toml
@@ -7,7 +7,8 @@ edition = "2021"
 
 [dependencies]
 tari_crypto = { version = "0.20.2", default-features = false }
-
+tari_common_types = { path = "../../../base_layer/common_types", version = "1.0.0-pre.16" }
+tari_common = { path = "../../../common" }
 ledger-transport = { git = "https://github.com/Zondax/ledger-rs", rev = "20e2a20" }
 ledger-transport-hid = { git = "https://github.com/Zondax/ledger-rs", rev = "20e2a20" }
 num-derive = "0.4.2"
